Adjust stance to be byte.

+++ cos-api/src/main/java/com/atlas/cos/attribute/CharacterAttributes.java
@@ -7,6 +7,6 @@ public record CharacterAttributes(Integer accountId, Integer worldId, String nam
                                   Integer intelligence, Integer hp, Integer mp, Integer maxHp, Integer maxMp, Integer meso,
                                   Integer hpMpUsed, Integer jobId, Integer skinColor, Byte gender, Integer fame, Integer hair,
                                   Integer face, Integer ap, String sp, Integer mapId, Integer spawnPoint, Integer gm, Integer x,
-                                  Integer y, Integer stance)
+                                  Integer y, Byte stance)
       implements AttributeResult {
 }
